singleton event filter
